Rename files "aahr-set" -> "multi-aahr".
